+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -17441,6 +17441,31 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         expanding_kwargs: dict,
         numeric_only: bool = False,
     ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_count_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_count_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    numeric_only=numeric_only,
+                )
+            )
+        qc = self._expanding_count_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            numeric_only=numeric_only,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_count_internal(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        numeric_only: bool = False,
+    ) -> "SnowflakeQueryCompiler":
         return self._window_agg(
             window_func=WindowFunction.EXPANDING,
             agg_func="count",
@@ -17449,6 +17474,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def expanding_sum(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_sum_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_sum_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_sum_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_sum_internal(
         self,
         fold_axis: Union[int, str],
         expanding_kwargs: dict,
@@ -17467,6 +17523,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def expanding_mean(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_mean_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_mean_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_mean_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_mean_internal(
         self,
         fold_axis: Union[int, str],
         expanding_kwargs: dict,
@@ -17503,6 +17590,40 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         engine: Optional[Literal["cython", "numba"]] = None,
         engine_kwargs: Optional[dict[str, bool]] = None,
     ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_var_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_var_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    ddof=ddof,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_var_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            ddof=ddof,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_var_internal(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        ddof: int = 1,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
         WarningMessage.warning_if_engine_args_is_set(
             "rolling_var", engine, engine_kwargs
         )
@@ -17514,6 +17635,40 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def expanding_std(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        ddof: int = 1,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_std_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_std_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    ddof=ddof,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_std_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            ddof=ddof,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_std_internal(
         self,
         fold_axis: Union[int, str],
         expanding_kwargs: dict,
@@ -17540,6 +17695,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         engine: Optional[Literal["cython", "numba"]] = None,
         engine_kwargs: Optional[dict[str, bool]] = None,
     ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_min_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_min_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_min_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_min_internal(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
         WarningMessage.warning_if_engine_args_is_set(
             "expanding_min", engine, engine_kwargs
         )
@@ -17551,6 +17737,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
 
     def expanding_max(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        numeric_only: bool = False,
+        engine: Optional[Literal["cython", "numba"]] = None,
+        engine_kwargs: Optional[dict[str, bool]] = None,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_max_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_max_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    numeric_only=numeric_only,
+                    engine=engine,
+                    engine_kwargs=engine_kwargs,
+                )
+            )
+        qc = self._expanding_max_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            numeric_only=numeric_only,
+            engine=engine,
+            engine_kwargs=engine_kwargs,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_max_internal(
         self,
         fold_axis: Union[int, str],
         expanding_kwargs: dict,
@@ -17640,6 +17857,34 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         ErrorMessage.method_not_implemented_error(name="quantile", class_="Expanding")
 
     def expanding_sem(
+        self,
+        fold_axis: Union[int, str],
+        expanding_kwargs: dict,
+        ddof: int = 1,
+        numeric_only: bool = False,
+    ) -> "SnowflakeQueryCompiler":
+        """
+        Wrapper around _expanding_sem_internal to be supported in faster pandas.
+        """
+        relaxed_query_compiler = None
+        if self._relaxed_query_compiler is not None:
+            relaxed_query_compiler = (
+                self._relaxed_query_compiler._expanding_sem_internal(
+                    fold_axis=fold_axis,
+                    expanding_kwargs=expanding_kwargs,
+                    ddof=ddof,
+                    numeric_only=numeric_only,
+                )
+            )
+        qc = self._expanding_sem_internal(
+            fold_axis=fold_axis,
+            expanding_kwargs=expanding_kwargs,
+            ddof=ddof,
+            numeric_only=numeric_only,
+        )
+        return self._maybe_set_relaxed_qc(qc, relaxed_query_compiler)
+
+    def _expanding_sem_internal(
         self,
         fold_axis: Union[int, str],
         expanding_kwargs: dict,