+++ CHANGELOG.md
@@ -83,6 +83,14 @@
   - `dt.month_name`
   - `dt.day_name`
   - `dt.strftime`  
+  - `expanding.min`
+  - `expanding.max`
+  - `expanding.count`
+  - `expanding.sum`
+  - `expanding.mean`
+  - `expanding.std`
+  - `expanding.var`
+  - `expanding.sem`
 - Make faster pandas disabled by default (opt-in instead of opt-out).
 
 ## 1.42.0 (2025-10-28)